+++ Imager/Imager.Server/Controllers/ImagerAppController.cs
@@ -11,19 +11,28 @@ namespace Imager.Server.Controllers;
 public class ImagerAppController : ControllerBase
 {
   private readonly IWebHostEnvironment _env;
+  private readonly ILogger<ImagerAppController> _logger;
   private readonly AzureBlobStorageService _blobService;
   private readonly AppDbContext _db;
 
-  public ImagerAppController(IWebHostEnvironment env, AzureBlobStorageService blobService, AppDbContext appDbContext)
+  public ImagerAppController(
+    IWebHostEnvironment env,
+    ILogger<ImagerAppController> logger,
+    AzureBlobStorageService blobService,
+    AppDbContext appDbContext)
   {
     _env = env;
+    _logger = logger;
     _blobService = blobService;
     _db = appDbContext;
   }
 
+  
+
   [HttpPost("upload")]
   public async Task<IActionResult> Upload([FromForm] ImageUploadRequest request)
   {
+    _logger.LogInformation("Upload endpoint invoked...");
     if (request.File == null || request.File.Length == 0) return BadRequest("No file uploaded.");
 
     // 1. Upload to Azure Blob (private container)
@@ -32,6 +41,7 @@ public class ImagerAppController : ControllerBase
     // 2. Save metadata to SQL
     var record = new ImageRecord
     {
+      ImageId = Guid.NewGuid(),
       Title = request.Title,
       Description = request.Description,
       BlobName = blobName,
@@ -44,16 +54,17 @@ public class ImagerAppController : ControllerBase
     await _db.SaveChangesAsync();
 
     return Ok(new {
-      id = record.Id,
+      id = record.ImageId, //record.Id,
       message = "Image uploaded and metadata saved"
     });
   }
 
   // Secure streaming endpoint (no public URL)
   [HttpGet("image/{id}")]
-  public async Task<IActionResult> GetImage(string id)
+  public async Task<IActionResult> GetImage(Guid id)
   {
-    var record = await _db.Images.FindAsync(id);
+    _logger.LogInformation("GetImage endpoint invoked...");
+    var record = _db.Images.Where(img => img.ImageId == id).ToList().FirstOrDefault();
     if (record == null) return NotFound();
 
     var stream = await _blobService.GetBlobStreamAsync(record.BlobName);