+++ Imager/Imager.Server/CHANGELOG.md
@@ -14,8 +14,13 @@ The following steps were used to generate this project:
 - Run `dotnet ef database update`
 - Run `dotnet ef migrations remove`
 
-Future Migraiton 
-- Run `dotnet ef migrations add AddTagsToImages`
+New EF Core Migration for Adding ImageId to ImageRecord
+- Modify `ImageRecord.cs` Id will be internal identifier and the ImageId is the external id to expose
+- Run `dotnet ef migrations add AddImageIdToImageRecord`
 - Run `dotnet ef database update`
+- Modified `ImagerAppController.cs` add ILogger<ImagerAppController> and changed GetImage {id} to be a GUID
+- Modify `Program.cs` make sure that CORS is working fix http to https://localhost:4200
 
-
+v1.0 Complete
+- The .NET API is able to save the image within Azure Blob storage and fetch from the location.
+  Along with saving the related meta data within a Azure SQL database using EF Core.